Frustum-cull small draws
This is an experiment, some games do a poor job of culling stuff, and
some transparent sprites can be very expensive if they cause a copy.
Skipping them if outside the viewport makes sense in that case.

One example are the flame sprites in #17797 .

Additionally, we should be able to cull through-mode draws easily, this
one doesn't even try.